[UPDATE] modulo login, se agrego datos de tipo Ususario, se agrega enpoint.

+++ src/main/java/org/edu/citas/controller/MontoController.java
@@ -183,7 +183,7 @@ public class MontoController {
             }
 
             response.put("response",1);
-            response.put("data", montos.byteValue());
+            response.put("data", montos);
 
             return  ResponseEntity.ok(response);
